Robotmk: Split remaining `robotmk` module into section and test plugin
CMK-14987
