+++ checkman/robotmk_test
@@ -1,4 +1,4 @@
-title: Robotmk
+title: Robotmk Test
 agents: robotmk
 catalog: robotframework
 license: GPLv2
